Cleaning up. Moved live implementation into it's own module.
