+++ Package.swift
@@ -12,7 +12,8 @@ let package = Package(
     .watchOS(.v9),
   ],
   products: [
-    .library(name: "SupabaseClient", targets: ["SupabaseClient"])
+    .library(name: "SupabaseClient", targets: ["SupabaseClient"]),
+    .library(name: "SupabaseClientLive", targets: ["SupabaseClientLive"])
   ],
   dependencies: [
     .package(
@@ -22,6 +23,14 @@ let package = Package(
     .package(
       url: "https://github.com/supabase-community/supabase-swift.git",
       from: "0.3.0"
+    ),
+    .package(
+      url: "https://github.com/supabase-community/gotrue-swift.git",
+      from: "1.0.0"
+    ),
+    .package(
+      url: "https://github.com/supabase-community/postgrest-swift.git",
+      from: "1.0.0"
     ),
     .package(
       url: "https://github.com/m-housh/swift-identified-storage.git",
@@ -38,12 +47,20 @@ let package = Package(
       dependencies: [
         .product(name: "Dependencies", package: "swift-dependencies"),
         .product(name: "IdentifiedStorage", package: "swift-identified-storage"),
+        .product(name: "GoTrue", package: "gotrue-swift"),
+        .product(name: "PostgREST", package: "postgrest-swift"),
+      ]
+    ),
+    .target(
+      name: "SupabaseClientLive",
+      dependencies: [
+        "SupabaseClient",
         .product(name: "Supabase", package: "supabase-swift"),
       ]
     ),
     .testTarget(
       name: "SupabaseClientTests",
-      dependencies: ["SupabaseClient"]
+      dependencies: ["SupabaseClient", "SupabaseClientLive"]
     ),
   ]
 )
