+++ Sources/SupabaseClient/AuthenticationError.swift
@@ -1,0 +1,3 @@
+import Foundation
+
+public struct AuthenticationError: Error { }
