+++ Sources/SupabaseClient/Mock/ClientAuth+Mock.swift
